Split APIHandler into single files
